+++ src/UserSearchContainer.js
@@ -8,7 +8,9 @@ import {
   StripesConnectedSource,
 } from '@folio/stripes/smart-components';
 
-import filterConfig from './filterConfig';
+import filterConfig, { filterConfigWithUserAssignedStatus } from './filterConfig';
+import { updateResourceData } from './utils';
+import { UAS } from './constants';
 
 const INITIAL_RESULT_COUNT = 30;
 const RESULT_COUNT_INCREMENT = 30;
@@ -32,6 +34,28 @@ const compileQuery = template(
   { interpolate: /%{([\s\S]+?)}/g }
 );
 
+export function buildQuery(queryParams, pathComponents, resourceData, logger, props) {
+  const filters = props.initialSelectedUsers ? filterConfigWithUserAssignedStatus : filterConfig;
+  const updatedResourceData = props.initialSelectedUsers && resourceData?.query?.filters?.substring(`${UAS}`) ? updateResourceData(resourceData) : resourceData;
+
+  return makeQueryFunction(
+    'cql.allRecords=1',
+    (parsedQuery, _, localProps) => localProps.query.query.trim().split(/\s+/).map(query => compileQuery({ query })).join(' and '),
+    {
+      // the keys in this object must match those passed to
+      // SearchAndSort's columnMapping prop
+      'active': 'active',
+      'name': 'personal.lastName personal.firstName',
+      'patronGroup': 'patronGroup.group',
+      'username': 'username',
+      'barcode': 'barcode',
+      'email': 'personal.email',
+    },
+    filters,
+    2,
+  )(queryParams, pathComponents, updatedResourceData, logger, props);
+}
+
 class UserSearchContainer extends React.Component {
   static manifest = Object.freeze({
     initializedFilterConfig: { initialValue: false },
@@ -46,24 +70,7 @@ class UserSearchContainer extends React.Component {
       perRequest: 100,
       path: 'users',
       GET: {
-        params: {
-          query: makeQueryFunction(
-            'cql.allRecords=1',
-            (parsedQuery, props, localProps) => localProps.query.query.trim().split(/\s+/).map(query => compileQuery({ query })).join(' and '),
-            {
-              // the keys in this object must match those passed to
-              // SearchAndSort's columnMapping prop
-              'active': 'active',
-              'name': 'personal.lastName personal.firstName',
-              'patronGroup': 'patronGroup.group',
-              'username': 'username',
-              'barcode': 'barcode',
-              'email': 'personal.email',
-            },
-            filterConfig,
-            2,
-          ),
-        },
+        params: { query: buildQuery },
         staticFallback: { params: {} },
       },
     },