+++ src/constants.js
@@ -1,0 +1,8 @@
+/* eslint-disable import/prefer-default-export */
+export const UAS = 'uas';
+export const ASSIGNED_FILTER_KEY = 'uas.Assigned';
+export const UNASSIGNED_FILTER_KEY = 'uas.Unassigned';
+export const ASSIGNED = 'Assigned';
+export const UNASSIGNED = 'Unassigned';
+export const ACTIVE = 'active';
+export const INACTIVE = 'inactive';